requirements_n.txt
flask==2.3.3
pymongo==4.5.0
pandas==2.0.3
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.6
sentence-transformers==2.2.2
chromadb==0.4.22
google-generativeai==0.3.2